% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{BSI}
\alias{BSI}
\title{BIS}
\usage{
BSI(lastprice, bid1, ask1, volume, AGGREGATE = FALSE)
}
\arguments{
\item{lastprice}{last trading price.}

\item{bid1}{previous orderbook's bid1 price.}

\item{ask1}{previous orderbook's ask1 price.}

\item{volume}{last trading volume.}

\item{AGGREGATE}{specyfing wether to aggretate all buyer/seller initiated
volumes together.}
}
\value{
a matrix of two columns corresponding to buyer and seller initialed
order flow.
}
\description{
market order flow.
}
\details{
extract market order flow form give transaction data.
}

