% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{cancelall}
\alias{cancelall}
\title{cancelall}
\usage{
cancelall(instrumentid = NULL, direction = NULL, action = NULL,
  pricemin = NULL, pricemax = NULL, orderid = NULL)
}
\arguments{
\item{instrumentid}{character, specifying a filter for instrument identifiers.}

\item{direction}{integer, specifying a filter for trading directions. 1 for long and -1 for short.}

\item{action}{character, specifying a filter for actions, can take value from one of "open","close","closetoday","closepreday"}

\item{pricemin}{numeric, specifying a filter for price lower limit.}

\item{pricemax}{numeric, specifying a filter for price upper limit.}

\item{orderid}{character, specifying the set of orderids to be canceled. NOTE: if orderid is not null, cancelall will disregard any other filters and cancel orders only by orderid}
}
\value{
nothing
}
\description{
cancel all satisfied orders
}
\examples{
\dontrun{
## cancel all orders satisfy direction==-1
cancelall(tradetime,direction==-1)
}
}
\author{
Chen
}
\seealso{
\link{replaceall}
}

