% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{capchange}
\alias{capchange}
\title{capchange}
\usage{
capchange(TODAY = stop("TODAY not supplied"),
  TOTAL = stop("TOTAL not supplied"), HANDS = stop("HANDS not supplied"),
  COMMISSION = stop("COMMISSION not supplied"))
}
\value{
nothing
}
\description{
capchange
}
\details{
capchange
}
\author{
Chen
}

