% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{CFEupdate}
\alias{CFEupdate}
\title{CFEupdate}
\usage{
CFEupdate(DATA = stop("DATA not supplied"),
  INSTRUMENTID = stop("INSTRUMENTID not supplied"))
}
\arguments{
\item{DATA}{data passed to strategy}

\item{INSTRUMENTID}{character, instrumment identifier, indicating currently traded instrument}

\item{INSTRUMENT}{environment name specified in initializeinstrument, type ?initializeinstrument for more information}
}
\value{
tradetime, lastprice, orderbook, preorderbook, volume
}
\description{
extract tradetime, lastprice, orderbook, preorderbook and volume from current data flow. update queuing orders and capital state.
}
\examples{
\dontrun{
initializestates(realtime=FALSE)

## generate an environment named 'TF', the instrument to be traded is 'TF1512'
initializeinstrument(instrument = "TF",instrumentid="TF1512",pbuyhands = seq(from = 32,by = 1,length.out = 5),
                     pbuyprice = seq(from = 22,by = 1,length.out = 5),
                     psellhands = seq(from = 37,by = 1,length.out = 5),
                     psellprice = seq(from = 27,by = 1,length.out = 5),
                     ptradetime = 2,plastprice = 4,pvolume = 12,
                     fee = c(long=0,short=0,closetoday=0,closepreday=0),
                     closeprior = "today",
                     timeformat = "\%Y-\%m-\%d \%H:\%M:\%S",
                     multiplier = 10000)

## strategy
TF1512 <- function(EXdata){
 CFEupdate(EXdata,TF,'TF1512')
 print(tradetime)
 print(lastprice)
 print(orderbook)
 print(preorderbook)
 print(volume)
}

## data srouce
datasource <- getHF_Future(instID = "TF1512", startDate = "2015-10-29", endDate = "2015-10-29")
## run strategy
for(i in 1:nrow(datasource)){TF1512(datasource[i,])}

}
}
\author{
Chen
}
\seealso{
\link{initializestates} \link{initializeinstrument}
}

