% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{chasecloseall}
\alias{chasecloseall}
\title{chasecloseall}
\usage{
chasecloseall(instrumentid, chasesleep = 1)
}
\arguments{
\item{instrumentid}{character, specyfing instrument to be closed.}

\item{chasesleep}{numeric, specyfing order chasing interval.}
}
\value{
nothing
}
\description{
chase close all holdings of a specific instrument.
}
\details{
chasecloseall can only close one instrument at a time, simulator
will recheck if the order price is equal to current bid1 or ask1 price every
chasesleep seconds, if not, simulator will cancel it and submit a new one.
This action will be repeated until all specified holdings are executed.
}
\seealso{
\link{closeall}
}

