% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{checklimit}
\alias{checklimit}
\title{checklimit}
\usage{
checklimit(instrumentdata, orderid)
}
\arguments{
\item{instrumentdata}{data.frame, generated from datasource}

\item{orderid}{character, specifying id of the limit order}
}
\value{
plot
}
\description{
checklimit
}
\details{
check limit order details
}
\author{
Chen
}

