% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{closeall}
\alias{closeall}
\title{closeall}
\usage{
closeall(instrumentid = "qtid", price = NULL, type = "limit")
}
\value{
nothing
}
\description{
closeall
}
\details{
closeall
}
\author{
Chen
}

