% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{closedprofit}
\alias{closedprofit}
\title{closed profit}
\usage{
closedprofit(instrumentid)
}
\arguments{
\item{instrumentid}{character, instrument identifier}
}
\value{
numeric, closed profit
}
\description{
calculate closed profit. require setting closed=TRUE in
HFTsimulator.
}
\details{
closed profit is the most recent cash when all holdings are
equal to zero. total_profit = holdings_profit + closed_profit.
}
\examples{
\dontrun{
## get closed profit of TF1603
closedprofit("TF1603")
}
}
\seealso{
\link{HFTsimulator} \link{holdingsprofit}
}

