% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{datamanipulation}
\alias{datamanipulation}
\title{datamanipulation}
\usage{
datamanipulation(instrumentdata, instrumentid)
}
\value{
formatted data
}
\description{
datamanipulation
}
\details{
datamanipulation
}
\author{
Chen
}

