% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{extractinfo}
\alias{extractinfo}
\title{extractinfo}
\usage{
extractinfo(info = c("tradetime", "lastprice", "volume", "orderbook", "HMOS",
  "presettleprice"), EXdata, ptradetime, plastprice, pvolume, pbuyhands,
  pbuyprice, psellhands, psellprice, ppresettleprice,
  timeformat = "\%Y-\%m-\%d \%H:\%M:\%OS")
}
\arguments{
\item{info}{information to be extracted from hf data, one of "tradetime","lastprice","volume","orderbook"}

\item{EXdata}{hf data}

\item{ptradetime}{position of tradetime}
}
\value{
information
}
\description{
extractinfo
}
\details{
extractinfo
}
\author{
Chen
}

