% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{furtherlazyfunctions}
\alias{furtherlazyfunctions}
\title{furtherlazyfunctions}
\usage{
furtherlazyfunctions()
}
\value{
nothing
}
\description{
furtherlazyfunctions
}
\details{
furtherlazyfunctions
}
\author{
Chen
}

