% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{HFTsimulator}
\alias{HFTsimulator}
\title{HFTsimulator}
\usage{
HFTsimulator(stg, ..., instrumentids, datalist, formatlist, tc = FALSE,
  Sleep = 1, DIGITSSECS = 3, septraded = FALSE, unclosed = TRUE,
  closed = TRUE, interdaily = FALSE, verboselimitpriors = TRUE)
}
\arguments{
\item{stg}{function, strategy function.}

\item{...}{parameters passed to stg.}

\item{instrumentids}{character, spefifying instruments to be traded.}

\item{datalist}{data.frame or list, specifying taq data used in the
simulation. datalist must be a list of data.frame(s) or a data.frame.}

\item{formatlist}{list, specifying taq data format, formatlist is either a
list of data format specifycation or a list of lists of specifications.}

\item{tc}{logical, indicating wehter to use a simulated tradecenter. when
tc=TRUE, only lazysubmission can be used as submit function in stg. Defalut
FALSE.}

\item{Sleep}{numeric, idle time length of simulated tradecenter, measured
in seconds, default 1.}

\item{DIGITSSECS}{integer, specifying second digits, default 3.}

\item{septraded}{logical, indicating wether to record traded orders
separately.}

\item{unclosed}{logical, indicating wether to track all unclosed orders,
set unclosed=TRUE when you need to calculate mean open price and open
profit. Default TRUE.}

\item{closed}{logical, indicating wether to track all zero holding states,
set closed=TRUE when you need to calculate close profit, default TRUE.}

\item{interdaily}{logical, indicating wether to support interdaily strategies,
most of the time MM strategies are appiled in intraday situations,
set it to TRUE only when you know exactly what you are doing. Defalut FALSE.}

\item{verboselimitpriors}{logical, indicating wether to record all prior 
limit orders' informations. if verboselimitpriors=TRUE,  simulator will
contatenate all limitpriors in a list named 'verbosepriors'. Default TRUE.}
}
\value{
a list containing all kinds of histories and current states.
}
\description{
high-frequency trading simulator.
}
\details{
Initialize simulator states, including simulation back ground
functionalities and many ohter simulator related parameters. All
states related variables are saved in an environment named
'.tradingstates'. Queuing orders and capital state will be saved and
kept updated in tradingstates during simulation. There are two improtant
data.frames stored in this envrionment, 'orders' and 'capital'. All
current queuing orders will be stored as one rows in orders during
simulation. if there is no queuing order, orders will be a data.frame
with 0 row. each instruments' capital state will be stored as one row in
capital. capital has at least one row. \code{queryorder()} and
\code{qureycapital()} can be used inside strategy function to fetch orders
and capital from .tradingstates.
}
\seealso{
\link{lazysubmission} \link{cancelall} \link{queryorder}
\link{querycapital} \link{meanopen} \link{holdingsprofit}
}

