% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{holdingsprofit}
\alias{holdingsprofit}
\title{holdings profit}
\usage{
holdingsprofit(instrumentid = character(), side = "long")
}
\arguments{
\item{instrumentid}{character, instrument identifier}

\item{side}{character, "long" or "short", specifying holdings's side}
}
\value{
holdings profit, numeric
}
\description{
calculate unclosed holdings' dynamic profit. require setting unclosed=TRUE in initializestates()
}
\details{
long holdings' dynamic profit = holdings * (lastprice - mean open price);    short holdings' dynamic profit = holdings * (mean open price - lastprice)
}
\examples{
\dontrun{
holdingsprofit("TF1603","long")
}
}
\author{
Chen
}
\seealso{
\link{initializestates} \link{meanopen} \link{closedprofit}
}

