% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{initializeENV}
\alias{initializeENV}
\title{initializeEnv}
\usage{
initializeENV(realtimeDATA, instruments, exchanges, multipliers, endofthedays,
  fee = c(long = 4e-06, short = 4e-06, closetoday = 4e-06, closepreday =
  4e-06), closeprior = "preday", tc = FALSE, Sleep = 1, IMLAZY = FALSE,
  DIGITSSECS = 3, STRINGSASFACTORS = FALSE, septraded = FALSE,
  unclosed = TRUE, closed = TRUE, interdaily = FALSE)
}
\description{
initialize all simulator states
}
\examples{
\dontrun{
instruments <- c("AL1603", "TF1606", "AU1606")
detail <- getdetail(instruments)
initializeENV(realtimeDATA = FALSE,ENVname = "AL",instruments = instruments[1],exchange = detail$exchanges[1],multiplier = detail$multipliers[1],endoftheday = detail$endofthedays[1])
initializeENV(realtimeDATA = FALSE,ENVname = "TF",instruments = instruments[2],exchange = detail$exchanges[2],multiplier = detail$multipliers[2],endoftheday = detail$endofthedays[2])
}
}

