% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{initializeinstrument}
\alias{initializeinstrument}
\title{initializeinstrument}
\usage{
initializeinstrument(instrumentid, pbuyhands, pbuyprice, psellhands, psellprice,
  ptradetime, plastprice, pvolume, ppresettleprice, fee = c(long = 0, short =
  0, closetoday = 0, closepreday = 0), closeprior = "today",
  timeformat = "\%Y\%m\%d\%H\%M\%OS", endoftheday = "15:15:00.000",
  multiplier = 10000)
}
\arguments{
\item{instrumentid}{character, instrument identifier, unique.}

\item{fee}{named numeric, specifying conmissions of different actions, including open, close, closetoday and closepreday. 'cost' in orderhistory and capitalhistory are result calculated by 'fee'.}

\item{closeprior}{character, specifying close priority when specified action='close' in ordersubmission. closeprior can only be one of 'today' and 'preday'. when closeprior='today', ordersubmission will close today's holdings prior than previous days', vise versa. type ?ordersubmission for more details.}

\item{timeformat}{character, specifying time format of the data source.}

\item{endoftheday}{character, specifying ending time of each trading day, simulator will move holdings to preholdings whenever new tradetime pass through 'endoftheday'.}

\item{multiplier}{numeric, quoted price * multiplier = real price.}

\item{instrument}{character, specifying the name of a generated environment, all details about the instruments interested will be stored in this environment. see 'Examples' below.}

\item{pXXX}{integer, specifying index of XXX in 'data source', see 'Details' and 'Examples' for more about 'data source'.}
}
\value{
an envrionment in .GlobalEnv containing all the informations specified above. the env's name is specified by parameter 'instrument'.
}
\description{
initialize instruments to be traded, including data structure, fee for different actions, data time formant, corresponding multipliers and initial holdings. if you have already set IMLAZY=TRUE in initializestates(), initializeinstrument() will generate plenty of additional objects in order to simplify strategy writting procedure,see 'Details' below.
}
\details{
IMLAZY: if IMLAZY=TRUE, initializeinstrument() will generate a bunch of expressions named INSTRUMENTID.ATTRIBUTE.STATE or INSTRUMENTID.ATTRIBUTE. for example, TF1603.longopen.non represents 'TF1603 has no long open order in queue' and TF1603.longopen represents 'all TF1603's long open orders', you can retrive prices of all the long open orders of TF1603 by typing TF1603.longopenX$X'price'. see 'Examples' for more information. part of the expressions are listed below:
         INSTRUMENTID.orders.non
         INSTRUMENTID.orders.exists
         INSTRUMENTID.longopen.exists
         INSTRUMENTID.longopen.non
         INSTRUMENTID.shortopen.exists
         INSTRUMENTID.shortopen.non
         INSTRUMENTID.longclose.exists
         INSTRUMENTID.longclose.non
         INSTRUMENTID.shortclose.exists
         INSTRUMENTID.shortclose.non
         INSTRUMENTID.longholdings.exists
         INSTRUMENTID.longholdings.non
         INSTRUMENTID.shortholdings.exists
         INSTRUMENTID.shortholdings.non
         INSTRUMENTID.holdings.exists
         INSTRUMENTID.holdings.non
         INSTRUMENTID.longopen
         INSTRUMENTID.shortopen
         INSTRUMENTID.longclose
         INSTRUMENTID.shortclose

data source: any kind of data sorce that can pass market data to strategy and simulator one row at a time.
}
\examples{
\dontrun{
## wirte capital and order histories to local file, don't sychronize target holdings, don't use simulated trade center and tell the simulator 'I'm lazy'.
initializestates(realtime=FALSE,writeholding=FALSE,tc=FALSE,IMLAZY = TRUE)

## generate an environment named 'TF', the instrument to be traded is 'TF1512'
initializeinstrument(instrument = "TF",instrumentid="TF1512",pbuyhands = seq(from = 32,by = 1,length.out = 5),
                     pbuyprice = seq(from = 22,by = 1,length.out = 5),
                     psellhands = seq(from = 37,by = 1,length.out = 5),
                     psellprice = seq(from = 27,by = 1,length.out = 5),
                     ptradetime = 2,plastprice = 4,pvolume = 12,
                     fee = c(long=0,short=0,closetoday=0,closepreday=0),
                     closeprior = "today",
                     timeformat = "\%Y-\%m-\%d \%H:\%M:\%S",
                     multiplier = 10000)

TF1512 <- function(EXdata){
 CFEupdate(EXdata,TF,'TF1512')

 ## IMLAZY=TRUE
 if(TF1512.holdings.non & TF1512.longopen.non)
   ordersubmission(tradetime,"TF1512",orderid="xxx",direction=1,price=0,hands=1,action='open')

 ## IMLAZY=FALSE, as a comparison
 ##mycapital <- querycapital(instrumentid="TF1512")
 ##myorders <- queryorders()
 ##if(mycapital$totallongholdings==0 & mycapital$totalshortholdings=0 & nrow(myorders[myorders$instrumentid=='TF1512'&myorders$direction==1&myorders$action=='open',])==0)
 ## ordersubmission(tradetime,"TF1512",orderid="xxx",direction=1,price=0,hands=1,action='open')
}

## data srouce
datasource <- getHF_Future(instID = "TF1512", startDate = "2015-10-29", endDate = "2015-10-29")
## backtest
for(i in 1:nrow(datasource)){TF1512(datasource[i,])}

}
}
\author{
Chen
}
\seealso{
\link{initializestates} \link{lazysubmission} \link{ordersubmission}
}

