% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{initializeTF}
\alias{initializeTF}
\title{initializeTF}
\usage{
initializeTF(realtimeDATA = TRUE, TFs = "TF1512", fee = c(long = 2.25e-06,
  short = 2.25e-06, closetoday = 2.25e-06, closepreday = 2.25e-06),
  closeprior = "today", tc = FALSE, Sleep = 1, IMLAZY = FALSE,
  DIGITSSECS = 3, STRINGSASFACTORS = FALSE, septraded = FALSE,
  unclosed = TRUE, closed = TRUE, interdaily = FALSE)
}
\arguments{
\item{realtimeDATA}{logical, indicating wether to use realtime data.}

\item{TFs}{character, specifying TF ids to be initialized.}

\item{fee}{named numeric, specifying conmissions of different actions, including open, close, closetoday and closepreday. 'cost' in orderhistory and capitalhistory are result calculated by 'fee'.}

\item{closeprior}{character, specifying close priority when specified action='close' in ordersubmission. closeprior can only be one of 'today' and 'preday'. when closeprior='today', ordersubmission will close today's holdings prior than previous days', vise versa. type ?ordersubmission for details.}

\item{tc}{logical, indicating wehter to use a simulated tradecenter. when tc=TRUE, submitmethod will be coerced to 'lazysubmission'(type ?lazysubmission for details). see 'Details' for more about tradecenter}

\item{Sleep}{numeric, idle time length of simulated tradecenter, measured in seconds, default 1. see 'Details' for more information.}

\item{IMLAZY}{logical, pleas set it to TRUE if you are lazy. type ?initializeinstrument for more infromation.}

\item{DIGITSSECS}{integer, second digits, default 3}

\item{septraded}{logical, indicating wether to save traded orders separately.}

\item{unclosed}{logical, indicating wether to track all unclosed orders, set unclosed=TRUE when you need to calculate mean open price and open profit. type ?meanopen for more infromation.}

\item{closed}{logical, indicating wether to track all zero holding states, set closed=TRUE when you need to calculate close profit.}

\item{interdaily}{logical, indicating wether to support interdaily trading.}
}
\value{
nothing
}
\description{
a simple wrapper of initializestates() and initializeinstrument() for Treasury Futures. type ?initializestates and ?initializeinstrument for details.
}
\author{
Chen
}
\seealso{
\link{initializestates} \link{initializeinstrument}
}

