% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{is.newday}
\alias{is.newday}
\title{is.newday}
\usage{
is.newday(instrumentid)
}
\arguments{
\item{instrumentid}{character, instrument identifier, unique.}
}
\value{
logical, indication wether current data come from a new trading day
}
\description{
is new day?
}
\details{
is new day?
}
\author{
Chen
}

