% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{isnewday}
\alias{isnewday}
\title{isnewday}
\usage{
isnewday(instrumentid)
}
\arguments{
\item{instrumentid}{character, instrument identifier, unique.}
}
\value{
logical, indication wether current data come from a new trading day.
}
\description{
check if current instrument's data comes from a new day.
}

