% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{lazysubmission}
\alias{lazysubmission}
\title{lazysubmission}
\usage{
lazysubmission(instrumentid, longholding = NULL, shortholding = NULL,
  tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\arguments{
\item{instrumentid}{character, instrument identifier}

\item{longholding}{integer, specifying target long holdings of 'instrumentid', longholding >=0}

\item{shortholding}{integer, specifying target short holdings of 'instrumentid', shortholding <= 0}
}
\value{
nothing
}
\description{
submit target holdings, trade center will cancel all irrevelant orders and chase bid1 or ask1 price automatically to achieve target holdings. this function can only be used when set tc=TRUE in initializestates()
}
\examples{
\dontrun{

 lazysubmission(tradetime,"TF1512",longholding=5,shortholding=-3)
}
}
\author{
Chen
}
\seealso{
\link{initializestates}
}

