% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{maxdrawdown}
\alias{maxdrawdown}
\title{maxdrawdown}
\usage{
maxdrawdown(pl, ddown)
}
\value{
a list containing draw-down range and value
}
\description{
maxdrawdown
}
\details{
maxdrawdown
}
\author{
Chen
}

