% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{orderchaser}
\alias{orderchaser}
\title{orderchaser}
\usage{
orderchaser(tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\description{
chase bid1 or ask1 price
}
\author{
Chen
}
\seealso{
\link{chasesubmission}
}

