% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{ordersubmission}
\alias{ordersubmission}
\title{ordersubmission}
\usage{
ordersubmission(instrumentid = "TF1603", orderid = NULL, direction = 1,
  price = 0, hands = 1, action = "open", timeoutlist = FALSE,
  timeoutchase = FALSE, timeoutsleep = 1, chaselist = FALSE,
  chasesleep = 1, tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\arguments{
\item{instrumentid}{character, instrument identifier}

\item{orderid}{character, specifying an unique order id, can be generated by randomid()}

\item{direction}{integer, specifying trading direction. 1 for long, -1 for short.}

\item{price}{numeric, specifiying order pirce.NOTE: when price=0, ordersubmission() will submit a market order; when price=NULL, ordersubmission() will take the corresponding bid1 or ask1 price as order price.}

\item{hands}{integer, specifying amount to be submitted.}

\item{action}{character, specifying submit action, action can take value from one of "open","close","closetoday","closepreday" and "cancel". amout submitted in action='close' can not be greater than the sum of current holdings and queuing open hands.}
}
\value{
order status
}
\description{
different kinds of order actions, including open, close, closetoday, closepreday and cancel.
}
\author{
Chen Haotian
}
\seealso{
\link{multisubmission} \link{timeoutchasesubmission} \link{timeoutsubmission} \link{chasesubmission}
}

