% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{perfectexecution}
\alias{perfectexecution}
\title{perfectexecution}
\usage{
perfectexecution(instrumentid, orderid = "xxx", direction, price, hands,
  action, type = "limit")
}
\value{
nothing
}
\description{
perfectexecution
}
\details{
perfectexecution
}
\author{
Chen
}

