% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{plottwocolumntable}
\alias{plottwocolumntable}
\title{plottwocolumntable}
\usage{
plottwocolumntable(d, h1 = -1, h2 = 1, v1 = 0, v2 = 0, s = 3)
}
\value{
plot
}
\description{
plottwocolumntable
}
\details{
plottwocolumntable
}
\author{
Chen
}

