% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{priororders}
\alias{priororders}
\title{priororders}
\usage{
priororders(mostrecentorderbook, orderid, direction, price)
}
\value{
nothing
}
\description{
priororders
}
\details{
priororders
}
\author{
Chen
}

