% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{querycapital}
\alias{querycapital}
\title{querycapital}
\usage{
querycapital(instrumentids = NULL)
}
\arguments{
\item{instrumentids}{specifying instrumentids to be queried, return total
capital status if instrumentids=NULL, default NULL.}
}
\value{
a data.frame of all specified instruments' current status
}
\description{
get current capital status.
}
\details{
return a data.frame of current capital status, each row
of the data.frame representing an instrument, if instrumentids is not NULL,
querycapital will return the capital status specified by instrumentids.
}
\examples{
\dontrun{
## get total capital status
querycapital()
## get capital status of TF1603 and T1603
querycapital(instrumentids=c("TF1603","T1603"))
}
}

