% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{queryorder}
\alias{queryorder}
\title{queryorder}
\usage{
queryorder(orderids = NULL)
}
\arguments{
\item{orderids}{specyfing order ids to be queried, return all orders if
orderids=NULL, default NULL.}
}
\value{
a data.frame of queuing orders.
}
\description{
get current queuing orders.
}
\details{
return a data.frame of current queuing orders, each row of
the data.frame representing an order, queryorder will return all of the
queuing orders if orderid is NULL. when there is no queuing orders,
queryorder will return a data.frame with 0 rows.
}
\examples{
\dontrun{
## get all queuing orders
queryorder()
## get all orders that satisfying orderid\%in\%c("order1","order2")
queryorder(orderids=c("order1","order2"))
}
}

