% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{randomid}
\alias{randomid}
\title{randomid}
\usage{
randomid(n)
}
\arguments{
\item{n}{number of chars}
}
\value{
character, order id
}
\description{
generage a random order id
}
\examples{
\dontrun{
## generate a 5 characters' order id
randomid(5)
}
}

