% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{submitmultilevelopen}
\alias{submitmultilevelopen}
\title{submitmultilevelopen}
\usage{
submitmultilevelopen(instrumentid, LEVELS = c(1, 2), hands = 1,
  cancelallother = FALSE, cancelprime = FALSE, cancelsub = FALSE,
  DIRECTION = 1, cancelnotinthebook = FALSE)
}
\arguments{
\item{instrumentid}{character, instrument identifier.}

\item{LEVELS}{integer, specifying postions in order book. Orders will be
submmited to these positions.}

\item{hands}{integer, specifying amount to be submitted.}

\item{cancelallother, }{logical, indicating wehter or not  cancel all other
orders that satisfying in the order book but with different prices.}

\item{cancelprime}{cancel all orders with higher priority price}

\item{cancelsub}{cancel all orders with lower priority price}

\item{DIRECTION}{integer, specifying trading direction. 1 for long,
-1 for short.}

\item{cancelnotinthebook}{cancel orders not in orderbook}
}
\value{
nothing.
}
\description{
submit multiple open orders while cancel all other orders
satisfying the cancel conditions, cancel conditions are specified by
cancelallother, cancelprime,  cancelsub and cancelnotinthebook.
}
\seealso{
\link{multisubmission} \link{cancelall}
}

