% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{sucker}
\alias{sucker}
\title{sucker}
\usage{
sucker(LONGHOLDINGS = stop("LONGHOLDINGS not supplied"),
  SHORTHOLDINGS = stop("SHORTHOLDINGS not supplied"))
}
\value{
nothing
}
\description{
sucker
}
\details{
sucker
}
\author{
Chen
}

