% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{summaryvpplot}
\alias{summaryvpplot}
\title{summaryvpplot}
\usage{
summaryvpplot(pd, pl, dd, ddinfo, lossinfo, wininfo, os, ss, pds, traded,
  SUMMARY = TRUE, TRADED = TRUE)
}
\value{
plot
}
\description{
summaryvpplot
}
\details{
summaryvpplot
}
\author{
Chen
}

