% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.r
\docType{data}
\name{TFformat}
\alias{TFformat}
\title{Data Format of Treasury Future's TAQ data}
\format{A list with 12 elements}
\usage{
TFformat
}
\description{
A dataset containing format informations. The variables are as follows:
}
\details{
\itemize{
   \item pbuyhands. position of bid volumes
   \item pbuyprice. position of bid prices
   \item psellhands. position of ask volumes
   \item psellprice. position of ask prices
   \item ptradetime. position of trade time
   \item plastprice. position of last price
   \item pvolume. position total volume
   \item fee. fee
   \item closeprior. close priority
   \item timeformat. time format
   \item multiplier. multiplier
}
}
\keyword{datasets}

