% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.r
\docType{data}
\name{TFtaq}
\alias{TFtaq}
\title{Treasury Future's TAQ data in 2015-12-25}
\format{A data frame with 19178 rows and 55 variables}
\usage{
TFtaq
}
\description{
A dataset containing all the TAQ informations. The variables are as follows:
}
\details{
\itemize{
   \item CONTRACTID. instrument id
   \item TDATETIME. trade time
   \item OPENPRICE. open price
   \item LASTPRICE. last price
   \item HIGHPRICE. high price
   \item LOWPRICE. low price
   \item SETTLEPRICE. settle price
   \item PRESETTLE. previous settle price
   \item CLOSEPRICE. close price
   \item PRECLOSE. previous close
   \item CQ. current trade volume
   \item VOLUME. total volume
   \item CM. cm
   \item AMOUNT. amount
   \item PREPOSITION. preposition
   \item POSITION. position
   \item POSITIONCHANGE. positionchange
   \item LIMITUP. limitup
   \item LIMITDOWN. limitdown
   \item SIDE. side
   \item OC. oc
   \item B01. bid price 01
   \item B02. bid price 02
   \item B03. bid price 03
   \item B04. bid price 04
   \item B05. bid price 05
   \item S01. ask price 01
   \item S02. ask price 02
   \item S03. ask price 03
   \item S04. ask price 04
   \item S05. ask price 05
   \item BV01. bid volume 01
   \item BV02. bid volume 02
   \item BV03. bid volume 03
   \item BV04. bid volume 04
   \item BV05. bid volume 05
   \item SV01. ask volume 01
   \item SV02. ask volume 02
   \item SV03. ask volume 03
   \item SV04. ask volume 04
   \item SV05. ask volume 05
   \item CURRDELTA. currdelta
   \item PREDELTA. predelta
   \item SETTLEMENTGROUPID. settlementgroupid
   \item SETTLEMENTID. settlementid
   \item CHANGE. change
   \item CHANGERATIO. changeratio
   \item CONTINUESIGN. continuesign
   \item TRADINGDATE. tradingdate
   \item LOCALTIME. localtime
   \item RECTIME. rectime
   \item EXCHANGECODE. exchange code
   \item ID. id
   \item UNIX. unix time stamp
   \item DATE. date, (%Y-%m-%d)
}
}
\keyword{datasets}

