% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{timeoutdetector}
\alias{timeoutdetector}
\title{timeoutdetector}
\usage{
timeoutdetector(tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\description{
detect timeout orders, ust be executed before orderchaser
}
\author{
Chen
}
\seealso{
\link{timeoutsubmission}
}

