% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{timeoutsubmission}
\alias{timeoutsubmission}
\title{timeoutsubmission}
\usage{
timeoutsubmission(instrumentid = "qtid", orderid = NULL, direction = 1,
  price = 0, hands = 1, action = "open", timeoutsleep = 1)
}
\arguments{
\item{instrumentid}{character, instrument identifier.}

\item{orderid}{character, specifying an unique order id, can be generated
by randomid().}

\item{direction}{integer, specifying trading direction. 1 for long,
-1 for short.}

\item{price}{numeric, specifiying order pirce.NOTE: when price=0,
ordersubmission() will submit a market order; when price=NULL,
ordersubmission() will take the corresponding bid1 or ask1 price as
submitted price.}

\item{hands}{integer, specifying amount to be submitted.}

\item{action}{character, specifying submit action, action can take value
from one of "open","close","closetoday","closepreday" and "cancel". amount
submitted in action='close' can not be greater than the sum of current
holdings and queuing open hands.}

\item{timeoutsleep}{numeric, specifying the timeout inverval in seconds.}
}
\value{
order status code.
}
\description{
submit an order with timeout checking. The order will be
canceled when it hasn't been executed for a duration longer than
timeoutsleep
}
\details{
timeoutsubmission is a wrapper of ordersubmission, it act the same
as ordersubmission(...,timeoutlist=TRUE,chaselist=FALSE)
}
\examples{
\dontrun{
## submit an open order, buy 1 hand of TF1603 at price 99
## cancel the order if it's not executed in the next 10 seconds
timeoutsubmission(instrumentid="TF1603",orderid=randomid(5),
                direction=1,price=99,hands=1,action="open".
                timeoutsleep=10)
}
}
\seealso{
\link{multisubmission} \link{timeoutchasesubmission} \link{ordersubmission} \link{chasesubmission}
}

