% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{tradecenter}
\alias{tradecenter}
\title{tradecenter}
\usage{
tradecenter(instrumentid,
  tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\value{
nothing
}
\description{
tradecenter
}
\details{
tradecenter
}
\author{
Chen
}

