% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{tradesummary}
\alias{tradesummary}
\title{tradesummary}
\usage{
tradesummary(instrumentdata, instrumentid = "qtid", limitorders = NULL,
  starttime = "09:15:00.000", endtime = "15:15:00.000", SUMMARY = TRUE,
  TRADED = TRUE)
}
\arguments{
\item{instrumentdata}{data.frame, generated from datasource}

\item{instrumentid}{character, specifying instrument id}

\item{starttime}{character, must be of format "%H:%M:%S"}

\item{endtime}{character, must be of format "%H:%M:%S"}
}
\value{
a named list and a plot
}
\description{
tradesummary
}
\details{
summary trade result
}
\author{
Chen
}

