% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{updatecapital}
\alias{updatecapital}
\title{updatecapital}
\usage{
updatecapital(instrumentid, direction, hands, action, tradeprice, fee,
  closeprior = "today", multiplier = 10000)
}
\value{
nothing
}
\description{
updatecapital
}
\details{
updatecapital
}
\author{
Chen
}

