% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{updateinstrument}
\alias{updateinstrument}
\title{updateinstrument}
\usage{
updateinstrument(instrumentid, lastprice, volume, orderbook, preorderbook, fee,
  closeprior = "today", multiplier)
}
\arguments{
\item{instrumentid}{instrumentid}

\item{lastprice}{lastprice}

\item{volume}{volume}

\item{orderbook}{orderbook}

\item{closeprior}{closeprior}

\item{tradetime}{tradetime}

\item{prelastprice}{prelastprice}

\item{key}{redis key}

\item{redisEnv}{redis environment}

\item{timeformat}{time format}
}
\value{
nothing
}
\description{
updateinstrument
}
\details{
updateinstrument
}
\author{
Chen
}

