% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{verboselimitpriors}
\alias{verboselimitpriors}
\title{verbose}
\usage{
verboselimitpriors(tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\description{
record all prior limit orders' informations. be careful when using this macro, simulator will contatenate all limitpriors in tradingstates to a gaint list named 'verbosepriors'
}

