% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{writecapitalhistory}
\alias{writecapitalhistory}
\title{writecapitalhistory}
\usage{
writecapitalhistory(instrumentid, tradeprice, tradehands, cost,
  tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\value{
nothing
}
\description{
writecapitalhistory
}
\details{
writecapitalhistory
}
\author{
Chen
}

