% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{writeorderhistory}
\alias{writeorderhistory}
\title{writeorderhistory}
\usage{
writeorderhistory(instrumentid, orderid, direction, hands, price, tradeprice,
  status, action, cost, tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\value{
nothing
}
\description{
writeorderhistory
}
\details{
writeorderhistory
}
\author{
Chen
}

