% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{closedprofit}
\alias{closedprofit}
\title{closed profit}
\usage{
closedprofit(instrumentid)
}
\arguments{
\item{instrumentid}{character, instrument identifier}
}
\value{
closed profit, numeric
}
\description{
calculate closed profit. require setting closed=TRUE in initializestates()
}
\details{
closed profit is the most recent cash value when all holdings are equal to zero
}
\examples{
\dontrun{
closedprofit("TF1603")
}
}
\author{
Chen
}
\seealso{
\link{initializestates} \link{holdingsprofit}
}

