% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{lazysubmission}
\alias{lazysubmission}
\title{lazysubmission}
\usage{
lazysubmission(instrumentid, longholding = NULL, shortholding = NULL)
}
\arguments{
\item{instrumentid}{character, instrument identifier}

\item{longholding}{integer, specifying target long holdings of
'instrumentid', longholding must be greater than or equal to 0.}

\item{shortholding}{integer, specifying target short holdings of
'instrumentid', shortholding must be less than or equal to 0.}
}
\value{
nothing
}
\description{
submit a target holding, simulator will cancel all irrevelant
orders and chase bid1 or ask1 price automatically until the target holding
}
\examples{
\dontrun{
 lazysubmission("TF1512",longholding=5,shortholding=-3)
}
}
\seealso{
\link{HFTsimulator}
}

