% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{multisubmission}
\alias{multisubmission}
\title{multisubmission}
\usage{
multisubmission(instrumentid = "qtid", orderid = NULL, direction = 1,
  price = NULL, hands = 1, action = "open", timeoutlist = FALSE,
  timeoutchase = FALSE, timeoutsleep = 1, chaselist = FALSE,
  chasesleep = 1)
}
\arguments{
\item{instrumentid}{character, instrument identifier}

\item{orderid}{character, if length(orderid)==0 (default), multisubmission will generate a random id for each order}

\item{direction}{integer, specifying trading direction. 1 for long, -1 for short.}

\item{price}{numeric, specifiying order pirce.default NULL. NOTE: when price=0, ordersubmission() will submit a market order; when price=NULL, ordersubmission() will take the corresponding bid1 or ask1 price as order price.}

\item{hands}{integer, specifying hands to be submitted.}

\item{action}{character, action can take value from one of "open","close","closetoday","closepreday" and "cancel". hands submitted in action='close' can not be greater than the sum of current holdings and queuing open hands.}
}
\value{
order status
}
\description{
submit multiple orders, a simple wrapper of ordersubmission(). instrumentid, direction, price, hands and action must be of length one or the same length with the number of orders; orderid must be of length zero or the same length with the number of orders!
}
\examples{
\dontrun{
## submit leng(orderbook$buybook$price[-1]) long open orders in TF1512.
multisubmission(instrumentid="TF1512",orderid=NULL,direction=1,price=orderbook$buybook$price[-1],hands=1,action='open')
}
}
\author{
Chen
}
\seealso{
\link{ordersubmission} \link{timeoutchasesubmission} \link{timeoutsubmission} \link{chasesubmission}
}

