% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{perfectexecution}
\alias{perfectexecution}
\title{perfectexecution}
\usage{
perfectexecution(instrumentid, orderid = "xxx", direction, price, hands,
  action)
}
\arguments{
\item{instrumentid}{character, instrument identifier.}

\item{orderid}{character, specifying an unique order id, can be generated
by randomid().}

\item{direction}{integer, specifying trading direction. 1 for long,
-1 for short.}

\item{price}{numeric, specifiying order pirce.}

\item{hands}{integer, specifying amount to be submitted.}

\item{action}{character, specifying submit action, action can take value
from one of "open","close","closetoday","closepreday" and "cancel". amount
submitted in action='close' can not be greater than the sum of current
holdings and queuing open hands.}
}
\value{
nothing.
}
\description{
execute and order immediatele with a specified price, without
going through the simulation system. Can be used to comparing simulated
strategy with a perfect situation.
}
\examples{
\dontrun{
## submit an open order, buy 1 hand of TF1603 at price 99
## the order will be executed immediately at price 99
perfectexecution(instrumentid="TF1603",orderid='xxx',
                direction=1,price=99,hands=1,action="open")
}
}

