% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{perfectexecution}
\alias{perfectexecution}
\title{perfectexecution}
\usage{
perfectexecution(instrumentid, orderid = "xxx", direction, price, hands,
  action, type = "limit",
  tradetime = .GlobalEnv$tradingstates$currenttradetime)
}
\value{
nothing
}
\description{
perfectexecution
}
\details{
perfectexecution
}
\author{
Chen
}

