% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{querycapital}
\alias{querycapital}
\title{querycapital}
\usage{
querycapital(instrumentids = NULL)
}
\arguments{
\item{instrumentids}{instrumentids to be queried, return all orders if instrumentids=NULL}
}
\value{
instruments
}
\description{
querycapital
rycapital
}
\details{
querycapital
}
\author{
Chen
}

