% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{queryorder}
\alias{queryorder}
\title{queryorder}
\usage{
queryorder(orderids = NULL)
}
\arguments{
\item{orderid}{orderid to be queried, return all orders if orderid=NULL}
}
\value{
order
}
\description{
queryorder
}
\details{
queryorder
}
\author{
Chen
}

