% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hftsimulator.r
\name{timeoutsubmission}
\alias{timeoutsubmission}
\title{timeoutsubmission}
\usage{
timeoutsubmission(instrumentid = "qtid", orderid = NULL, direction = 1,
  price = 0, hands = 1, action = "open", timeoutsleep = 1)
}
\arguments{
\item{instrumentid}{character, instrument identifier}

\item{orderid}{character, unique order id, can be generated by randomid()}

\item{direction}{integer, specifying trading direction. 1 for long, -1 for short.}

\item{price}{numeric, specifiying order pirce.NOTE: when price=0, ordersubmission() will submit a market order; when price=NULL, ordersubmission() will take the corresponding bid1 or ask1 price as order price.}

\item{hands}{integer, specifying hands to be submitted.}

\item{action}{character, action can take value from one of "open","close","closetoday","closepreday" and "cancel". hands submitted in action='close' can not be greater than the sum of current holdings and queuing open hands.}

\item{timeoutsleep}{numeric, specifying time out length, seconds}
}
\value{
order status
}
\description{
submit an order with timeout checking. The order will be canceled when it hasn't been executed for a duration longer than 'timeoutsleep'
}
\author{
Chen
}
\seealso{
\link{multisubmission} \link{timeoutchasesubmission} \link{ordersubmission} \link{chasesubmission}
}

